Fixes broken build
