+++ .github/workflows/build.yaml
@@ -49,4 +49,4 @@ jobs:
         run: ./gradlew spotlessCheck
 
       - name: Run Unit Tests
-        run: ./gradlew testBookdiaryDebugUnitTest
+        run: ./gradlew testStandardDebugUnitTest
